implement invalid ops on DTA/PA
